Allow Scrapy & Django containers to access the db
Adds a PostgreSQL container to the Docker Compose stack and configure
both containers to access the same database. Within Scrapy container
we're using Peewee as an ORM because of its simplicity. The pro is that
its API is simple and straighforward, allowing no migrations (what
should invite developers to handle migrations in Django, which is a good
thing). The con is that there are to models.py to maintain, one in
Django and one in Scrapy.